[REFACTO] computer + company

+++ src/app/computer/computer-details/computer-details.component.ts
@@ -22,34 +22,38 @@ export class ComputerDetailsComponent implements OnInit {
     this.computerService.getById(this.id)
       .subscribe(computer => this.computer = computer,
         error =>  {
-          if (error.status == 404) {
+          if (error.status === 404) {
             this.snackBar.open('This page does not exist.', 'Close', {
               panelClass: 'snackbar-error',
-              duration: 2500,});
+              duration: 2500});
           } else if (error.status > 499) {
             this.snackBar.open('Can\'t reach the database. Press F5 or try later.', 'Close', {
               panelClass: 'snackbar-error',
-              duration: 2500,});
+              duration: 2500});
           } else {
             this.snackBar.open(error.status + ': Press F5 or try later.', 'Close', {
               panelClass: 'snackbar-error',
-              duration: 2500,});
+              duration: 2500});
           }
           this.router.navigate([`/computers/`]);
       } );
   }
 
   delete(computer: Computer) {
-    this.computerService.deleteComputer(computer.id).subscribe(
-      () => {
-        this.snackBar.open('Computer "' + computer.name + '" deleted.', 'close', {
-                panelClass: 'snackbar-info',
-                duration: 2500});
-      }, error =>  {
-        this.snackBar.open('Can\'t delete the computer. Try again.', 'close', {
-          panelClass: 'snackbar-error',
-          duration: 2500});
-      } );
+    if (confirm('Are you sure to delete this computer ?')) {
+      this.computerService.deleteComputer(computer.id).subscribe(
+        () => {
+          this.snackBar.open('Computer "' + computer.name + ' deleted.', 'close', {
+            panelClass: 'snackbar-info',
+            duration: 2500
+          });
+        }, error => {
+          this.snackBar.open('Can\'t delete the computer. Try again.', 'close', {
+            panelClass: 'snackbar-error',
+            duration: 2500
+          });
+        });
+    }
   }
 
 }
